requirements.txt
streamlit==0.69.2
matplotlib==3.2.2
numpy==1.18.5
seaborn==0.10.1
pandas==1.0.5
Pillow==8.0.1
scikit_learn==0.23.2
